requirements.txt
bitcoin==1.1.42
certifi==2017.11.5
cffi==1.11.2
chardet==3.0.4
click==6.7
cytoolz==0.8.2
Django==1.11.7
djangorestframework==3.7.3
eth-abi==0.5.0
eth-keyfile==0.4.0
eth-keys==0.1.0b3
eth-tester==0.1.0b3
eth-testrpc==1.3.3
eth-utils==0.7.1
ethereum==1.6.1
ethjsonrpc==0.3.0
idna==2.6
json-rpc==1.10.8
pbkdf2==1.3
pkg-resources==0.0.0
psycopg2==2.7.3.2
py-solc==2.0.0
pycparser==2.18
pycryptodome==3.4.7
pyethash==0.1.27
pylru==1.0.9
pysha3==1.0.2
pytz==2017.3
PyYAML==3.12
repoze.lru==0.7
requests==2.9.1
rlp==0.6.0
scrypt==0.8.0
secp256k1==0.13.2
semantic-version==2.6.0
toolz==0.8.2
typing==3.6.2
urllib3==1.22
web3==3.16.4
Werkzeug==0.12.2
